+++ service/src/main/kotlin/fi/espoo/evaka/vasu/VasuTemplateController.kt
@@ -5,12 +5,15 @@
 package fi.espoo.evaka.vasu
 
 import fi.espoo.evaka.Audit
+import fi.espoo.evaka.shared.VasuTemplateId
 import fi.espoo.evaka.shared.auth.AuthenticatedUser
 import fi.espoo.evaka.shared.auth.UserRole
 import fi.espoo.evaka.shared.db.Database
 import fi.espoo.evaka.shared.domain.Conflict
 import fi.espoo.evaka.shared.domain.FiniteDateRange
 import fi.espoo.evaka.shared.domain.NotFound
+import fi.espoo.evaka.shared.security.AccessControl
+import fi.espoo.evaka.shared.security.Action
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
@@ -20,11 +23,12 @@ import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
-import java.util.UUID
 
 @RestController
 @RequestMapping("/vasu/templates")
-class VasuTemplateController {
+class VasuTemplateController(
+    private val accessControl: AccessControl
+) {
     data class CreateTemplateRequest(
         val name: String,
         val valid: FiniteDateRange,
@@ -36,9 +40,9 @@ class VasuTemplateController {
         db: Database.Connection,
         user: AuthenticatedUser,
         @RequestBody body: CreateTemplateRequest
-    ): UUID {
+    ): VasuTemplateId {
         Audit.VasuTemplateCreate.log()
-        user.requireOneOfRoles(UserRole.ADMIN)
+        accessControl.requirePermissionFor(user, Action.Global.CREATE_VASU_TEMPLATE)
 
         return db.transaction {
             it.insertVasuTemplate(
@@ -54,11 +58,11 @@ class VasuTemplateController {
     fun editTemplate(
         db: Database.Connection,
         user: AuthenticatedUser,
-        @PathVariable id: UUID,
+        @PathVariable id: VasuTemplateId,
         @RequestBody body: VasuTemplateUpdate
     ) {
         Audit.VasuTemplateEdit.log()
-        user.requireOneOfRoles(UserRole.ADMIN)
+        accessControl.requirePermissionFor(user, Action.VasuTemplate.UPDATE, id)
 
         db.transaction { tx ->
             val template = tx.getVasuTemplateForUpdate(id) ?: throw NotFound("Template not found")
@@ -76,11 +80,11 @@ class VasuTemplateController {
     fun copyTemplate(
         db: Database.Connection,
         user: AuthenticatedUser,
-        @PathVariable id: UUID,
+        @PathVariable id: VasuTemplateId,
         @RequestBody body: CopyTemplateRequest
-    ): UUID {
+    ): VasuTemplateId {
         Audit.VasuTemplateCopy.log(id)
-        user.requireOneOfRoles(UserRole.ADMIN)
+        accessControl.requirePermissionFor(user, Action.VasuTemplate.COPY, id)
 
         return db.transaction {
             val template = it.getVasuTemplate(id) ?: throw NotFound("template not found")
@@ -100,12 +104,7 @@ class VasuTemplateController {
         @RequestParam(required = false) validOnly: Boolean = false
     ): List<VasuTemplateSummary> {
         Audit.VasuTemplateRead.log()
-        user.requireOneOfRoles(
-            UserRole.ADMIN,
-            UserRole.UNIT_SUPERVISOR,
-            UserRole.SPECIAL_EDUCATION_TEACHER,
-            UserRole.STAFF
-        )
+        accessControl.requirePermissionFor(user, Action.Global.READ_VASU_TEMPLATE)
 
         return db.read { tx -> tx.getVasuTemplates(validOnly) }
     }
@@ -114,10 +113,11 @@ class VasuTemplateController {
     fun getTemplate(
         db: Database.Connection,
         user: AuthenticatedUser,
-        @PathVariable id: UUID
+        @PathVariable id: VasuTemplateId
     ): VasuTemplate {
         Audit.VasuTemplateRead.log(id)
         user.requireOneOfRoles(UserRole.ADMIN)
+        accessControl.requirePermissionFor(user, Action.VasuTemplate.READ, id)
 
         return db.read { tx -> tx.getVasuTemplate(id) } ?: throw NotFound("template $id not found")
     }
@@ -126,10 +126,10 @@ class VasuTemplateController {
     fun deleteTemplate(
         db: Database.Connection,
         user: AuthenticatedUser,
-        @PathVariable id: UUID
+        @PathVariable id: VasuTemplateId
     ) {
         Audit.VasuTemplateDelete.log(id)
-        user.requireOneOfRoles(UserRole.ADMIN)
+        accessControl.requirePermissionFor(user, Action.VasuTemplate.DELETE, id)
 
         db.transaction { it.deleteUnusedVasuTemplate(id) }
     }
@@ -138,11 +138,11 @@ class VasuTemplateController {
     fun putTemplateContent(
         db: Database.Connection,
         user: AuthenticatedUser,
-        @PathVariable id: UUID,
+        @PathVariable id: VasuTemplateId,
         @RequestBody content: VasuContent
     ) {
         Audit.VasuTemplateUpdate.log(id)
-        user.requireOneOfRoles(UserRole.ADMIN)
+        accessControl.requirePermissionFor(user, Action.VasuTemplate.UPDATE, id)
 
         db.transaction { tx ->
             val template = tx.getVasuTemplateForUpdate(id) ?: throw NotFound("template $id not found")
